# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HelmetWash API",
        "description": "REST API HelmetWash yang sinkron dengan model layanan, outlet, pesanan, dan riwayat status pada database.",
        "contact": {
            "name": "HelmetWash Support",
            "email": "admin@helmetwash.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "HelmetWash API Server"
        }
    ],
    "paths": {
...
        "/api/admin/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Semua pesanan (Admin)",
                "description": "Mengambil seluruh pesanan yang tersimpan di tabel orders beserta riwayat status.",
                "operationId": "getAllOrders",
                "responses": {
                    "200": {
                        "description": "Daftar pesanan berhasil didapat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/orders/{code}/status": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update status pesanan (Admin)",
                "description": "Memperbarui status pesanan dan mencatat perubahan ke tabel order_status_histories.",
                "operationId": "updateOrderStatus",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Kode pesanan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ORD-12345"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "Dicuci",
                                        "enum": [
                                            "Diterima",
                                            "Dicuci",
                                            "Selesai",
                                            "Diantar"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status pesanan diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Status pesanan diperbarui"
                                        },
                                        "order": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "ORD-12345"
                    },
                    "customerName": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "phone": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "serviceId": {
                        "type": "string",
                        "example": "basic"
                    },
                    "serviceName": {
                        "type": "string",
                        "example": "Paket Basic"
                    },
                    "addons": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "antijamur"
                        }
                    },
                    "deliveryId": {
                        "type": "string",
                        "example": "pickup"
                    },
                    "deliveryName": {
                        "type": "string",
                        "example": "Pick-up & Delivery"
                    },
                    "address": {
                        "type": "string",
                        "example": "Jl. Melati No. 5"
                    },
                    "priceBase": {
                        "type": "integer",
                        "example": 25000
                    },
                    "priceAddons": {
                        "type": "integer",
                        "example": 5000
                    },
                    "priceDelivery": {
                        "type": "integer",
                        "example": 10000
                    },
                    "totalPrice": {
                        "type": "integer",
                        "example": 40000
                    },
                    "status": {
                        "type": "string",
                        "example": "Diterima"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-02T10:30:00.000Z"
                    },
                    "payment": {
                        "properties": {
                            "method": {
                                "type": "string",
                                "example": "qris"
                            },
                            "status": {
                                "type": "string",
                                "example": "paid"
                            }
                        },
                        "type": "object"
                    },
                    "history": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "Diterima"
                                },
                                "date": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2026-07-02T10:30:00.000Z"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT yang didapat dari POST /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}